+++ src/mock/v1.2/intent-b.ts
@@ -1,8 +1,8 @@
 import { closeWindowOnCompletion, onFdc3Ready } from "./mock-functions";
 import { DesktopAgent } from "fdc3_1_2/dist/api/DesktopAgent";
 import { sendContextToTests } from "../v1.2/mock-functions";
-import { AppControlContext } from "../../test/common/common-types";
-import { Intent } from "../../test/v2.0/advanced/intent-support-2.0";
+import { AppControlContext } from "../../context-types";
+import { Intent } from "../../test/v2.0/support/intent-support-2.0";
 declare let fdc3: DesktopAgent;
 
 onFdc3Ready().then(async () => {
