+++ src/mock/v2.0/intent-d.ts
@@ -1,6 +1,6 @@
 import { closeWindowOnCompletion, onFdc3Ready } from "./mock-functions";
 import { DesktopAgent } from "fdc3_2_0";
-import { Intent } from "../../test/v2.0/advanced/intent-support-2.0";
+import { Intent } from "../../test/v2.0/support/intent-support-2.0";
 declare let fdc3: DesktopAgent;
 
 onFdc3Ready().then(async () => {
