+++ src/mock/v2.0/intent-f.ts
@@ -1,7 +1,7 @@
 import { closeWindowOnCompletion, onFdc3Ready, sendContextToTests, validateContext } from "./mock-functions";
 import { DesktopAgent } from "fdc3_2_0";
 import { wait } from "../../utils";
-import { ContextType, Intent } from "../../test/v2.0/advanced/intent-support-2.0";
+import { ContextType, Intent } from "../../test/v2.0/support/intent-support-2.0";
 declare let fdc3: DesktopAgent;
 
 //used in '2.0-RaiseIntentPrivateChannelResult'
