+++ src/mock/v2.0/metadata.ts
@@ -1,7 +1,7 @@
 import { closeWindowOnCompletion, onFdc3Ready, sendContextToTests } from "./mock-functions";
 import { DesktopAgent } from "fdc3_2_0";
 import { MetadataAppCommandContext } from "../../test/v2.0/advanced/fdc3.findInstances";
-import { Intent } from "../../test/v2.0/advanced/intent-support-2.0";
+import { Intent } from "../../test/v2.0/support/intent-support-2.0";
 declare let fdc3: DesktopAgent;
 
 onFdc3Ready().then(async () => {
