+++ src/test/v1.2/support/channels-support-1.2.ts
@@ -2,8 +2,8 @@ import { assert, expect } from "chai";
 import { Channel, Context, DesktopAgent, Listener } from "fdc3_1_2";
 import constants from "../../../constants";
 import { wait } from "../../../utils";
-import { ChannelControl, ChannelsAppConfig, ChannelsAppContext } from "../../common/channel-control";
-import { AppControlContext } from "../../common/common-types";
+import { ChannelControl, ChannelsAppConfig, ChannelsAppContext } from "../../common/control/channel-control";
+import { AppControlContext } from "../../../context-types";
 
 declare let fdc3: DesktopAgent;
