+++ src/test/v1.2/support/intent-support-1.2.ts
