+++ src/test/v2.0/advanced/fdc3.open.ts
@@ -1,8 +1,8 @@
 import constants from "../../../constants";
 import { getCommonOpenTests } from "../../common/fdc3.open";
-import { openApp, OpenCommonConfig } from "../../common/open-control";
+import { openApp, OpenCommonConfig } from "../../common/control/open-control";
 import { APIDocumentation2_0 } from "../apiDocuments-2.0";
-import { expectAppTimeoutErrorOnOpen, OpenControl2_0 } from "./open-support-2.0";
+import { expectAppTimeoutErrorOnOpen, OpenControl2_0 } from "../support/open-support-2.0";
 
 const openDocs = "\r\nDocumentation: " + APIDocumentation2_0 + "\r\nCause:";
 const control = new OpenControl2_0();
