+++ src/test/v2.0/advanced/fdc3.raiseIntent-Result.ts
@@ -1,4 +1,4 @@
-import { RaiseIntentControl2_0, IntentResultType, IntentApp, ContextType, Intent } from "./intent-support-2.0";
+import { RaiseIntentControl2_0, IntentResultType, IntentApp, ContextType, Intent } from "../support/intent-support-2.0";
 
 const control = new RaiseIntentControl2_0();
